Struct inexor_rgf_core_model::component::Component

pub struct Component {
    pub ty: ComponentTypeId,
    pub description: String,
    pub properties: Vec<PropertyType>,
    pub extensions: Vec<Extension>,
}

A component defines a set of properties to be applied to entity types and relation types.

Fields
ty: ComponentTypeId

The type definition of the component.

description: String

Textual description of the component.

properties: Vec<PropertyType>

The properties which are applied on entity or relation instances.

extensions: Vec<Extension>

Component specific extensions

Implementations

impl Component

pub fn new<T: Into<ComponentTypeId>, S: Into<String>>( ty: T, description: S, properties: Vec<PropertyType>, extensions: Vec<Extension> ) -> Component

pub fn new_from_type<S: Into<String>>( namespace: S, type_name: S, description: S, properties: Vec<PropertyType>, extensions: Vec<Extension> ) -> Component

Constructs a new component with the given name and properties

pub fn new_without_extensions<T: Into<ComponentTypeId>, S: Into<String>>( ty: T, description: S, properties: Vec<PropertyType> ) -> Component

Constructs a new component with the given name and properties

pub fn new_without_properties<T: Into<ComponentTypeId>, S: Into<String>>( ty: T, description: S, extensions: Vec<Extension> ) -> Component

Constructs an component with the given name but without properties

pub fn has_property<S: Into<String>>(&self, property_name: S) -> bool

Returns true, if the component contains a property with the given name.

pub fn has_extension(&self, ty: &ExtensionTypeId) -> bool

Returns true, if the component contains an extension with the given type.