Struct inexor_rgf_core_model::reactive_flow_instance::ReactiveFlowInstance

pub struct ReactiveFlowInstance {
    pub id: Uuid,
    pub ty: EntityTypeId,
    pub entity_instances: RwLock<HashMap<Uuid, Arc<ReactiveEntityInstance>>>,
    pub relation_instances: RwLock<HashMap<EdgeKey, Arc<ReactiveRelationInstance>>>,
    pub entities_added: RwLock<Vec<Uuid>>,
    pub entities_removed: RwLock<Vec<Uuid>>,
    pub relations_added: RwLock<Vec<EdgeKey>>,
    pub relations_removed: RwLock<Vec<EdgeKey>>,
}

Fields
id: Uuid

The id of the flow corresponds to the id of the wrapper entity instance.

ty: EntityTypeId

The type definition of the entity type of the wrapper entity instance.

entity_instances: RwLock<HashMap<Uuid, Arc<ReactiveEntityInstance>>>

The flow contains entity instances. The entity instance may also be contained in other flows.

relation_instances: RwLock<HashMap<EdgeKey, Arc<ReactiveRelationInstance>>>

The flow contains relation instances. The relation instances may also be contained in other flows.

entities_added: RwLock<Vec<Uuid>>

List of entities that has been added since creation of the flow.

entities_removed: RwLock<Vec<Uuid>>

List of entities that has been removed since creation of the flow.

relations_added: RwLock<Vec<EdgeKey>>

List of relations that has been added since creation of the flow.

relations_removed: RwLock<Vec<EdgeKey>>

List of relations that has been removed since creation of the flow.

Implementations

impl ReactiveFlowInstance

pub fn new( wrapper_entity_instance: Arc<ReactiveEntityInstance> ) -> ReactiveFlowInstance

pub fn has_entity(&self, entity_instance: Arc<ReactiveEntityInstance>) -> bool

pub fn has_entity_by_id(&self, id: Uuid) -> bool

pub fn get_entity(&self, id: Uuid) -> Option<Arc<ReactiveEntityInstance>>

pub fn get_wrapper_entity_instance(&self) -> Option<Arc<ReactiveEntityInstance>>

pub fn add_entity(&self, entity_instance: Arc<ReactiveEntityInstance>)

pub fn remove_entity(&self, id: Uuid)

pub fn has_relation( &self, relation_instance: Arc<ReactiveRelationInstance> ) -> bool

pub fn has_relation_by_key(&self, edge_key: &EdgeKey) -> bool

pub fn get_relation( &self, edge_key: &EdgeKey ) -> Option<Arc<ReactiveRelationInstance>>

pub fn add_relation(&self, relation_instance: Arc<ReactiveRelationInstance>)

pub fn remove_relation(&self, edge_key: &EdgeKey)

pub fn tick(&self)