Struct inexor_rgf_core_model::reactive_relation_instance::ReactiveRelationInstance

pub struct ReactiveRelationInstance {
    pub outbound: Arc<ReactiveEntityInstance>,
    pub ty: RelationInstanceTypeId,
    pub inbound: Arc<ReactiveEntityInstance>,
    pub description: String,
    pub properties: DashMap<String, ReactivePropertyInstance>,
    pub components: DashSet<ComponentTypeId>,
    pub behaviours: DashSet<BehaviourTypeId>,
}

Reactive instance of a relation in the directed property graph.

Property Graph: The relation instance can store properties.

Directed Graph: The direction of the relation point from the outbound entity instance to the inbound entity instance.

Reactive Instance: The properties are streams with a local copies of the last result. The streams can be connected, combined, folded or zipped.

One example for a directed reactive relation instance is a connector which propagates changes on a property of the outbound entity to a property of the inbound entity.

Another example would be the velocity transformation which are also using the streams of the properties of the outbound entity, the inbound entity and/or the relation itself.

Last but not least relation instances can be used for semantic representations like the current camera of a player: Player–(CurrentCamera)–>Camera

Fields
outbound: Arc<ReactiveEntityInstance>

The outbound entity instance.

ty: RelationInstanceTypeId

The type definition of the relation type.

inbound: Arc<ReactiveEntityInstance>

The outbound entity instance.

description: String

An optional description of the relation instance.

properties: DashMap<String, ReactivePropertyInstance>

The reactive properties.

components: DashSet<ComponentTypeId>

The names of the components which are applied on this relation instance.

behaviours: DashSet<BehaviourTypeId>

The names of the behaviours which are applied on this relation instance.

Implementations

impl ReactiveRelationInstance

pub fn new_from_properties( outbound: Arc<ReactiveEntityInstance>, inbound: Arc<ReactiveEntityInstance>, properties: EdgeProperties ) -> Result<ReactiveRelationInstance, ()>

pub fn new_from_instance( outbound: Arc<ReactiveEntityInstance>, inbound: Arc<ReactiveEntityInstance>, instance: RelationInstance ) -> ReactiveRelationInstance

pub fn new_from_type_with_properties<S: Into<String>>( namespace: S, outbound: Arc<ReactiveEntityInstance>, type_name: S, inbound: Arc<ReactiveEntityInstance>, properties: HashMap<String, Value> ) -> ReactiveRelationInstance

pub fn relation_type_id(&self) -> RelationTypeId

Returns the inner relation type id.

pub fn instance_id(&self) -> String

Returns the relation instance type id.

pub fn get_key(&self) -> EdgeKey