Struct inexor_rgf_core_model::relation_instance::RelationInstance

pub struct RelationInstance {
    pub outbound_id: Uuid,
    pub ty: RelationInstanceTypeId,
    pub inbound_id: Uuid,
    pub description: String,
    pub properties: HashMap<String, Value>,
    pub extensions: Vec<Extension>,
}

Relation instances are edges from an outbound entity instance to an inbound entity instance.

The relation instance is of a relation type. The relation type defines the entity types of the outbound entity instance and the inbound entity instance. Furthermore the relation type defines which properties (name, data type, socket type) a relation instance have to have.

In contrast to the relation type, the relation instance stores values/ documents in it’s properties.

Fields
outbound_id: Uuid

The id of the outbound vertex.

ty: RelationInstanceTypeId

The type definition of the relation instance type.

inbound_id: Uuid

The id of the inbound vertex.

description: String

Textual description of the relation instance.

properties: HashMap<String, Value>

The properties of then relation instance.

Each property is represented by it’s name (String) and it’s value. The value is a representation of a JSON. Therefore the value can be boolean, number, string, array or an object. For more information about the data types please look at https://docs.serde.rs/serde_json/value/enum.Value.html

extensions: Vec<Extension>

Relation instance specific extensions.

Implementations

impl RelationInstance

pub fn new<T: Into<RelationInstanceTypeId>>( outbound_id: Uuid, ty: T, inbound_id: Uuid, properties: HashMap<String, Value> ) -> RelationInstance

Constructs a new relation instance with the given outbound_id, type, inbound_id and properties

pub fn new_from_type_unique_id<S: Into<String>>( namespace: S, outbound_id: Uuid, type_name: S, inbound_id: Uuid, properties: HashMap<String, Value> ) -> RelationInstance

Constructs a new relation instance with the given outbound_id, type, inbound_id and properties

pub fn new_from_type_unique_for_instance_id<S: Into<String>>( namespace: S, outbound_id: Uuid, type_name: S, instance_id: S, inbound_id: Uuid, properties: HashMap<String, Value> ) -> RelationInstance

Constructs a new relation instance with the given outbound_id, type, inbound_id and properties

pub fn new_from_type_with_random_instance_id<S: Into<String>>( namespace: S, outbound_id: Uuid, type_name: S, inbound_id: Uuid, properties: HashMap<String, Value> ) -> RelationInstance

Constructs a new relation instance with the given outbound_id, type, inbound_id and properties

pub fn new_without_properties<T: Into<RelationInstanceTypeId>>( outbound_id: Uuid, ty: T, inbound_id: Uuid ) -> RelationInstance

Constructs a new relation instance with the given outbound_id, type, inbound_id but without properties

pub fn relation_type_id(&self) -> RelationTypeId

Returns the inner relation type id.

pub fn instance_id(&self) -> String

Returns the relation instance type id.

pub fn get_key(&self) -> EdgeKey

Returns the edge key of the relation instance.